refactor: Seed and reset functions
